+++ e2e-tests/development-runtime/gatsby-node.js
@@ -15,7 +15,7 @@ exports.onCreateNode = function onCreateNode({
 
       createNodeField({
         name: `slug`,
-        value: slug.replace("/", ""),
+        value: slug,
         node,
       })
       break
@@ -59,7 +59,7 @@ exports.createPages = async function createPages({
   data.posts.edges.forEach(({ node }) => {
     const { slug } = node.fields
     createPage({
-      path: `/${slug}`,
+      path: slug,
       component: blogPostTemplate,
       context: {
         slug: slug,