+++ packages/gatsby-plugin-utils/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./validate"
+export * from "./joi"